refactor: extract TaskRow into shared core/ui component
