+++ app/src/main/java/com/example/twdist_android/core/ui/components/task/TaskRow.kt
@@ -1,12 +1,12 @@
-package com.example.twdist_android.features.today.presentation.components
+package com.example.twdist_android.core.ui.components.task
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Card
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.RadioButton
@@ -17,12 +17,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.twdist_android.features.today.presentation.model.TaskState
 
 @Composable
-fun TodayTaskRow(
-    task: TaskState,
-    onCompleted: (TaskState) -> Unit,
+fun TaskRow(
+    task: TaskRowState,
+    onCompleted: (TaskRowState) -> Unit,
     modifier: Modifier = Modifier
 ) {
     Card(
@@ -56,14 +55,14 @@ fun TodayTaskRow(
 
 @Preview(showBackground = true)
 @Composable
-fun TodayTaskRowPreview() {
-    TodayTaskRow(
-        task = TaskState(
-            title = "Write API integration tests",
-            projectName = "Backend",
+private fun TaskRowPreview() {
+    TaskRow(
+        task = TaskRowState(
             id = 0L,
             projectId = 1L,
-            sectionId = 1L
+            sectionId = 1L,
+            title = "Write API integration tests",
+            projectName = "Backend"
         ),
         onCompleted = {}
     )
