+++ app/src/main/java/com/example/twdist_android/features/today/presentation/model/TodayUiEvent.kt
@@ -1,5 +1,7 @@
 package com.example.twdist_android.features.today.presentation.model
 
+import com.example.twdist_android.core.ui.components.task.TaskRowState
+
 sealed interface TodayUiEvent {
-    data class TaskCompleted(val task: TaskState) : TodayUiEvent
+    data class TaskCompleted(val task: TaskRowState) : TodayUiEvent
 }
